Add files via upload
Abstract—Automatic recognition of emotions is one of the most challenging of tasks. A sophisticated learning algorithm, which may represent high-level abstraction, is required to detect emotions from EEG signals. The detection of emotions from EEG signals requires a sophisticated learning algorithm that can reect high-level abstraction. Different techniques have been applied to reinforce the robustness of the emotion recognition systems using electroencephalogram (EEG) signals, in particular the learning of spatiotemporal traits. In this approach, the use of a model inspired from Siamese Network which deals with data through concatenation and 3D convolution. In this approach for training Adam optimizer was used. This model is a custom model, which can deal with a stack for a single time instant, which gives better efﬁciency than 3D Convolutional Network. This type of model is usually used for Face Recognition, but use of such a model for this purpose opens a new horizon of opportunity for Automatic recognition of emotion. Highest accuracy of our project is around 70% has been achieved by Dense Net and lowest accuracy was 30%. Keywords—3D Convolutional Neural Network, Seed IV, Siamese Network, Adam Optimizer,DenseNet, Deep Learning.